src/models/requirements.txt
google-cloud-storage
pandas
scikit-learn
joblib
numpy